move worked?

+++ shared_data/include/shared_data.h
@@ -94,7 +94,7 @@ extern SharedDataMutex<bool> motor_action; // Command to start motor action / si
 extern SharedDataMutex<bool> calculate_corrections; // Signal for main process to correct motor movement durring linear travel
 extern SharedDataMutex<float> left_motor_set_speed; // Speed factor (-1 -> 1)
 extern SharedDataMutex<float> right_motor_set_speed;
-extern SharedDataMutex<int64_t> left_encoder_count_shared;
-extern SharedDataMutex<int64_t> right_encoder_count_shared;
+extern SharedDataMutex<int32_t> left_encoder_count_shared;
+extern SharedDataMutex<int32_t> right_encoder_count_shared;
 
 #endif
